fix:  reimplementation of canvas and editor stores with flow management and UI state providers
