+++ apps/web/src/components/editor/flowEditor/editorToolbox.tsx
@@ -1,15 +1,18 @@
+"use client";
 import { Stack } from "@mantine/core";
 import React from "react";
 import AddNewBlockButton from "./toolbox/addNewBlockButton";
 import AddStickyNoteButton from "./toolbox/addStickyNoteBlockButton";
 import AiButton from "./toolbox/aiButton";
+import { useFlowEditorContext } from "./flowEditorContext";
 
 const EditorToolbox = () => {
+  const { features } = useFlowEditorContext();
   return (
     <Stack align="center">
-      <AddNewBlockButton />
-      <AddStickyNoteButton />
-      <AiButton />
+      {features.addBlockSidebar && <AddNewBlockButton />}
+      {features.stickyNoteBtn && <AddStickyNoteButton />}
+      {features.aiAssistant && <AiButton />}
     </Stack>
   );
 };
