+++ apps/web/src/components/editor/flowEditor/toolbox/addStickyNoteBlockButton.tsx
